refactor: extract review routes into build_review_router
